Pulling the high water detector types into their own files
