+++ src/Marten/Events/Daemon/HighWaterDetector.cs
@@ -8,16 +8,7 @@ using Npgsql;
 
 namespace Marten.Events.Daemon
 {
-    internal class HighWaterStatistics
-    {
-        public long LastMark { get; set; }
-        public long HighestSequence { get; set; }
-        public long CurrentMark { get; set; }
-        public bool HasChanged => CurrentMark > LastMark;
-        public DateTimeOffset? LastUpdated { get; set; }
-    }
-
-    internal class HighWaterDetector
+    internal class HighWaterDetector: IHighWaterDetector
     {
         private readonly ITenant _tenant;
         private readonly NpgsqlCommand _gapDetection;
