+++ src/Marten/Events/Daemon/IHighWaterDetector.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Marten.Events.Daemon
+{
+    internal interface IHighWaterDetector
+    {
+        Task<HighWaterStatistics> DetectInSafeZone(DateTimeOffset safeTimestamp, CancellationToken token);
+        Task<HighWaterStatistics> Detect(CancellationToken token);
+    }
+}
